# Reduce code duplication in banded_broadcast