`_banded_broadcast!` methods for `BandedColumn` are considerably large functions, and separating out the common parts and not inlining it actually helps with performance.
On master
```julia
julia> A = BandedMatrix(0=>Float64[1:30000;]); C = copy(A);

julia> @btime $C .= $A .* 2;
  81.493 μs (0 allocations: 0 bytes)
```
This PR
```julia
julia> @btime $C .= $A .* 2;
  75.632 μs (2 allocations: 48 bytes)
```
The two allocations arise because the function isn't inlined, but overall this seems pretty insignificant.

This also helps with TTFX:
On master
```julia
julia> A = BandedMatrix(0=>Float64[1:3000;]); C = copy(A);

julia> @time C .= A .* 2;
  0.620778 seconds (893.19 k allocations: 59.337 MiB, 4.86% gc time, 99.90% compilation time)

julia> @time C .= 2 .* A;
  0.535827 seconds (845.23 k allocations: 56.215 MiB, 3.08% gc time, 99.93% compilation time)
```
This PR
```julia
julia> @time C .= A .* 2;
  0.280671 seconds (778.59 k allocations: 53.679 MiB, 10.94% gc time, 99.84% compilation time)

julia> @time C .= 2 .* A;
  0.244315 seconds (751.66 k allocations: 51.822 MiB, 6.71% gc time, 99.81% compilation time)
```
Something similar is also done for the non-`BandedColumn` layout methods, although these need to be inlined for performance. In this case, the TTFX goes up slightly, but performance improves 
On master
```julia
julia> @time C' .= 2 .* A';
  0.131595 seconds (170.85 k allocations: 11.632 MiB, 7.74% gc time, 99.65% compilation time)

julia> @time C' .= A' .* 2;
  0.100251 seconds (100.30 k allocations: 6.801 MiB, 99.57% compilation time)

julia> @btime $C' .= $A' .* 2;
  24.479 μs (0 allocations: 0 bytes)

julia> @btime $C' .= 2 .* $A';
  24.453 μs (0 allocations: 0 bytes)
```
This PR
```julia
julia> @time C' .= 2 .* A';
  0.151288 seconds (198.19 k allocations: 13.432 MiB, 99.66% compilation time)

julia> @time C' .= A' .* 2;
  0.122727 seconds (127.62 k allocations: 8.610 MiB, 99.63% compilation time)

julia> @btime $C' .= $A' .* 2;
  18.426 μs (0 allocations: 0 bytes)

julia> @btime $C' .= 2 .* $A';
  18.382 μs (0 allocations: 0 bytes)
```